refactor: TableGroup의 유효성 검사를 OrderTables에서 수행하도록 합니다.
